HTBHF-1965 replace start page acceptance tests
